Minor refactor
